Tablas actualizadas

+++ db/mysql-init/02-init-tables.sql
@@ -1,7 +1,7 @@
 USE keyly;
 
 CREATE TABLE `Sucursals` (
-  `id` INT NOT NULL AUTO_INCREMENT,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
   `nom` VARCHAR(255) NOT NULL,
   `direccio` VARCHAR(255) NOT NULL,
   `ciutat` VARCHAR(255) NOT NULL,
@@ -12,32 +12,32 @@ CREATE TABLE `Sucursals` (
 );
 
 CREATE TABLE `Departament` (
-  `id` INT NOT NULL AUTO_INCREMENT,
-  `sucursal_id` INT NOT NULL,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `sucursal_id` BIGINT NOT NULL,
   `nom` VARCHAR(255) NOT NULL,
   PRIMARY KEY (`id`),
   CONSTRAINT `fk_departament_sucursal` FOREIGN KEY (`sucursal_id`) REFERENCES `Sucursals` (`id`)
 );
 
 CREATE TABLE `Rols` (
-  `id` INT NOT NULL AUTO_INCREMENT,
-  `sucursal_id` INT NOT NULL,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `sucursal_id` BIGINT NOT NULL,
   `nom` VARCHAR(255) NOT NULL,
   PRIMARY KEY (`id`),
   CONSTRAINT `fk_rol_sucursal` FOREIGN KEY (`sucursal_id`) REFERENCES `Sucursals` (`id`)
 );
 
 CREATE TABLE `Dominis` (
-  `id` INT NOT NULL AUTO_INCREMENT,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
   `nom` VARCHAR(255) NOT NULL,
   PRIMARY KEY (`id`)
 );
 
 CREATE TABLE `Usuaris` (
-  `id` INT NOT NULL AUTO_INCREMENT,
-  `sucursal_id` INT NOT NULL,
-  `departament_id` INT NOT NULL,
-  `rol_id` INT NOT NULL,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `sucursal_id` BIGINT NOT NULL,
+  `departament_id` BIGINT NOT NULL,
+  `rol_id` BIGINT NOT NULL,
   `nom` VARCHAR(255) NOT NULL,
   `correu` VARCHAR(255) NOT NULL UNIQUE,
   `contrasenya_master` VARCHAR(60) NOT NULL,
@@ -52,8 +52,8 @@ CREATE TABLE `Usuaris` (
 );
 
 CREATE TABLE `Bagul` (
-  `id` INT NOT NULL AUTO_INCREMENT,
-  `usuari_id` INT NOT NULL,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `usuari_id` BIGINT NOT NULL,
   `nom` VARCHAR(255) NOT NULL,
   `correu` VARCHAR(255) NOT NULL,
   `imatge` VARCHAR(255) NULL,
@@ -67,8 +67,8 @@ CREATE TABLE `Bagul` (
 );
 
 CREATE TABLE `Favorits` (
-  `id` INT NOT NULL AUTO_INCREMENT,
-  `usuari_id` INT NOT NULL,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `usuari_id` BIGINT NOT NULL,
   `item_id` INT NOT NULL,
   PRIMARY KEY (`id`),
   CONSTRAINT `fk_favorits_usuari` FOREIGN KEY (`usuari_id`) REFERENCES `Usuaris` (`id`),
@@ -76,8 +76,8 @@ CREATE TABLE `Favorits` (
 );
 
 CREATE TABLE `Carpetes` (
-  `id` INT NOT NULL AUTO_INCREMENT,
-  `propietari_id` INT NOT NULL,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `propietari_id` BIGINT NOT NULL,
   `nom` VARCHAR(255) NOT NULL,
   `data_creacio` DATE NOT NULL,
   PRIMARY KEY (`id`),
@@ -85,18 +85,18 @@ CREATE TABLE `Carpetes` (
 );
 
 CREATE TABLE `Carpetes_Items` (
-  `id` INT NOT NULL AUTO_INCREMENT,
-  `carpeta_id` INT NOT NULL,
-  `item_id` INT NOT NULL,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `carpeta_id` BIGINT NOT NULL,
+  `item_id` BIGINT NOT NULL,
   PRIMARY KEY (`id`),
   CONSTRAINT `fk_carpetes_items_item` FOREIGN KEY (`item_id`) REFERENCES `Bagul` (`id`),
   CONSTRAINT `fk_carpetes_items_carpeta` FOREIGN KEY (`carpeta_id`) REFERENCES `Carpetes` (`id`)
 );
 
 CREATE TABLE `Carpetes_Compartides` (
-  `id` INT NOT NULL AUTO_INCREMENT,
-  `carpeta_id` INT NOT NULL,
-  `usuari_id` INT NOT NULL,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `carpeta_id` BIGINT NOT NULL,
+  `usuari_id` BIGINT NOT NULL,
   `permisos` ENUM('lectura','escritura','administrador') NOT NULL,
   PRIMARY KEY (`id`),
   CONSTRAINT `fk_carpetes_compartides_carpeta` FOREIGN KEY (`carpeta_id`) REFERENCES `Carpetes` (`id`),
